64-bit timestamp

+++ app/inc/rtt_logger.h
@@ -14,7 +14,7 @@ class RttLogger: public Logger<RttLogger> {
   Rtt::Manager& m_manager;
   Rtt::CobsWriter m_writer;
 
-  void startMessage(uint32_t timestamp);
+  void startMessage(uint64_t timestamp);
   void appendData(const uint8_t* data, uint32_t length);
   void appendString(const char* str);
   void finishMessage();
